+++ chain-spec-generator/src/main.rs
@@ -77,8 +77,16 @@ fn main() -> Result<(), String> {
 				Box::new(system_parachains_specs::encointer_kusama_local_testnet_config) as Box<_>,
 			),
 			(
+				"coretime-kusama",
+				Box::new(system_parachains_specs::coretime_kusama_config) as Box<_>,
+			),
+			(
 				"coretime-kusama-local",
 				Box::new(system_parachains_specs::coretime_kusama_local_testnet_config) as Box<_>,
+			),
+			(
+				"coretime-polkadot",
+				Box::new(system_parachains_specs::coretime_polkadot_config) as Box<_>,
 			),
 			(
 				"coretime-polkadot-local",